Rename the path from embeddedwebview into webview Refactor the webviewclient and embeddedwebviewauthorizationstrategy
