--- common/src/main/java/com/microsoft/identity/common/internal/ui/embeddedwebview/challengehandlers/ChallengeFactory.java
@@ -1,65 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// All rights reserved.
-//
-// This code is licensed under the MIT License.
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files(the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions :
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-package com.microsoft.identity.common.internal.ui.embeddedwebview.challengehandlers;
-
-import android.support.annotation.NonNull;
-import android.webkit.HttpAuthHandler;
-import android.webkit.WebView;
-
-import com.microsoft.identity.common.exception.ClientException;
-
-/**
- * Factory capable of producing different kinds of challenges.
- */
-
-public final class ChallengeFactory {
-    /**
-     * Factory method to get new NtlmChallenge object.
-     *
-     * @param view    the WebView that is initiating the callback
-     * @param handler the HttpAuthHandler used to set the WebView's response
-     * @param host    the host requiring authentication
-     * @param realm   the realm for which authentication is required
-     * @return NtlmChallenge
-     */
-    public static NtlmChallenge getNtlmChallenge(final WebView view,
-                                                 final HttpAuthHandler handler,
-                                                 final String host,
-                                                 final String realm) {
-        return new NtlmChallenge(view, handler, host, realm);
-    }
-
-    /**
-     * Factory method to get new PKeyAuthChallenge object.
-     *
-     * @param redirectUri Location: urn:http-auth:CertAuth?Nonce=<noncevalue>
-     *                    &CertAuthorities=<distinguished names of CAs>&Version=1.0
-     *                    &SubmitUrl=<URL to submit response>&Context=<server state that
-     *                    client must convey back>
-     * @return PKeyAuthChallenge
-     * @throws ClientException if the validation of redirectUri fails.
-     */
-    public static PKeyAuthChallenge getPKeyAuthChallenge(@NonNull final String redirectUri) throws ClientException {
-        return new PKeyAuthChallenge(redirectUri);
-    }
-}
