--- common/src/main/java/com/microsoft/identity/common/internal/ui/embeddedwebview/challengehandlers/NtlmChallenge.java
@@ -1,59 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// All rights reserved.
-//
-// This code is licensed under the MIT License.
-//
-// Permission is hereby granted, free of charge, to any person obtaining a copy
-// of this software and associated documentation files(the "Software"), to deal
-// in the Software without restriction, including without limitation the rights
-// to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
-// copies of the Software, and to permit persons to whom the Software is
-// furnished to do so, subject to the following conditions :
-//
-// The above copyright notice and this permission notice shall be included in
-// all copies or substantial portions of the Software.
-//
-// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
-// THE SOFTWARE.
-package com.microsoft.identity.common.internal.ui.embeddedwebview.challengehandlers;
-
-import android.webkit.HttpAuthHandler;
-import android.webkit.WebView;
-
-public class NtlmChallenge {
-    private HttpAuthHandler mHandler;
-    private WebView mView;
-    private String mHost;
-    private String mRealm;
-
-    NtlmChallenge(final WebView view,
-                  final HttpAuthHandler handler,
-                  final String host,
-                  final String realm) {
-        mHandler = handler;
-        mView = view;
-        mHost = host;
-        mRealm = realm;
-    }
-
-    HttpAuthHandler getHandler() {
-        return mHandler;
-    }
-
-    WebView getView() {
-        return mView;
-    }
-
-    String getHost() {
-        return mHost;
-    }
-
-    String getRealm() {
-        return mRealm;
-    }
-}
