+++ common/src/main/java/com/microsoft/identity/common/internal/ui/webview/EmbeddedWebViewAuthorizationStrategy.java
@@ -20,7 +20,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.internal.ui.embeddedwebview;
+package com.microsoft.identity.common.internal.ui.webview;
 
 import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
@@ -33,6 +33,7 @@ import android.view.View;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
 
+import com.microsoft.identity.common.R;
 import com.microsoft.identity.common.adal.internal.AuthenticationConstants;
 import com.microsoft.identity.common.adal.internal.util.StringExtensions;
 import com.microsoft.identity.common.exception.ClientException;
@@ -41,6 +42,7 @@ import com.microsoft.identity.common.internal.providers.oauth2.AuthorizationRequ
 import com.microsoft.identity.common.internal.providers.oauth2.AuthorizationResult;
 import com.microsoft.identity.common.internal.providers.oauth2.AuthorizationResultFuture;
 import com.microsoft.identity.common.internal.providers.oauth2.AuthorizationStrategy;
+import com.microsoft.identity.common.internal.ui.webview.challengehandlers.IChallengeCompletionCallback;
 
 import java.io.UnsupportedEncodingException;
 import java.util.concurrent.Future;
@@ -48,28 +50,23 @@ import java.util.concurrent.Future;
 /**
  * Serve as a class to do the OAuth2 auth code grant flow with Android embedded web view.
  */
-public class EmbeddedWebViewAuthorizationStrategy <GenericWebViewClient extends OAuth2WebViewClient,
-        GenericAuthorizationRequest extends AuthorizationRequest,
+public class EmbeddedWebViewAuthorizationStrategy <GenericAuthorizationRequest extends AuthorizationRequest,
         GenericAuthorizationResult extends AuthorizationResult>
         extends AuthorizationStrategy<GenericAuthorizationRequest, GenericAuthorizationResult> {
 
     private static final String TAG = StringExtensions.class.getSimpleName();
+    private Activity mActivity;
+    private IChallengeCompletionCallback mCallback;
     private WebView mWebView;
-    private String mStartUrl;
     private AuthorizationResultFuture mFuture;
 
     /**
      * Constructor of EmbeddedWebViewAuthorizationStrategy.
-     *
-     * @param webViewClient Generic WebViewClient
-     * @throws UnsupportedEncodingException thrown when the Character Encoding is not supported
-     * @throws ClientException throw when error happens during the authorization
      */
-    public EmbeddedWebViewAuthorizationStrategy(final GenericWebViewClient webViewClient, final WebView webView)
-            throws UnsupportedEncodingException, ClientException {
-        //TODO validate auth request in OAuth2Strategy.
-        setUpWebView(webViewClient, webView);
-        mStartUrl = webViewClient.getRequest().getAuthorizationStartUrl();
+    public EmbeddedWebViewAuthorizationStrategy(@NonNull final Activity activity,
+                                                @NonNull IChallengeCompletionCallback callback) {
+        mActivity = activity;
+        mCallback = callback;
     }
 
     /**
@@ -77,7 +74,7 @@ public class EmbeddedWebViewAuthorizationStrategy <GenericWebViewClient extends 
      * @param webViewClient AzureActiveDirectoryWebViewClient
      */
     @SuppressLint({"SetJavaScriptEnabled", "ClickableViewAccessibility"})
-    private void setUpWebView(final GenericWebViewClient webViewClient, final WebView webView) {
+    private void setUpWebView(final AzureActiveDirectoryWebViewClient webViewClient, final WebView webView) {
         // Create the Web View to show the page
         mWebView = webView;
         WebSettings userAgentSetting = mWebView.getSettings();
@@ -110,27 +107,33 @@ public class EmbeddedWebViewAuthorizationStrategy <GenericWebViewClient extends 
     /**
      * Load the start url for auth grant flow. It will load the black page first to avoid error for not loading web view.
      */
-    private void loadStartUrl() {
+    private void loadStartUrl(final String startUrl) {
         mWebView.post(new Runnable() {
             @Override
             public void run() {
                 // load blank first to avoid error for not loading webview
                 mWebView.loadUrl("about:blank");
                 Logger.verbose(TAG, "Launching embedded WebView for acquiring auth code.");
-                Logger.verbosePII(TAG, "The start url is" + mStartUrl);
-                mWebView.loadUrl(mStartUrl);
+                Logger.verbosePII(TAG, "The start url is" + startUrl);
+                mWebView.loadUrl(startUrl);
             }
         });
     }
 
-    @Override
-    public Future<AuthorizationResult> requestAuthorization(AuthorizationRequest request) throws UnsupportedEncodingException {
+    public void requestAuthorization(@Nullable final GenericAuthorizationRequest request) throws UnsupportedEncodingException, ClientException {
+        Logger.verbose(TAG, "Perform the authorization request with embedded webView.");
 
+        final AzureActiveDirectoryWebViewClient webViewClient
+                = new AzureActiveDirectoryWebViewClient(mActivity, mCallback, request.getRedirectUri());
+        final WebView webView = mActivity.findViewById(R.id.webview);
+        setUpWebView(webViewClient, webView);
+        loadStartUrl(request.getAuthorizationStartUrl());
+        // requestAuthorization could not return the authorization result
+        // The activity result is set in AuthenticationActivity.setResult()
+        // And AuthenticationActivity in ADAL/MSAL is not moved into common
 
         Logger.verbose(TAG, "Perform the authorization request with embedded webView.");
-        loadStartUrl();
-        mFuture = new AuthorizationResultFuture();
-        return mFuture;
+        loadStartUrl(request.getAuthorizationStartUrl());
     }
 
     @Override
