+++ common/src/main/java/com/microsoft/identity/common/internal/ui/webview/challengehandlers/NtlmChallengeHandler.java
@@ -20,7 +20,7 @@
 // LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
-package com.microsoft.identity.common.internal.ui.embeddedwebview.challengehandlers;
+package com.microsoft.identity.common.internal.ui.webview.challengehandlers;
 
 import android.app.Activity;
 import android.app.AlertDialog;
